backend/pyproject.toml
[project]
name = "backend-v2"
version = "0.1.0"
description = "V2 backend of quibble"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.3.6",
    "dj-database-url>=3.0.1",
    "django>=5.2.6",
    "django-base16-admin>=0.1.5",
    "django-cors-headers>=4.7.0",
    "django-environ>=0.12.0",
    "django-ltree-2>=0.1.12",
    "django-ninja>=1.4.3",
    "django-redis>=6.0.0",
    "django-resized>=1.0.3",
    "django-shortuuid>=0.1.2",
    "django-storages[s3]>=1.14.6",
    "google-genai>=0.2.0",
    "httpx>=0.28.1",
    "pgvector>=0.4.2",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "sentence-transformers>=5.2.2",
    "whitenoise[brotli]>=6.11.0",
]

[dependency-groups]
dev = [
    "django-stubs>=5.2.7",
]
